Extract some common logic to use cases
