+++ app/src/main/java/io/github/wulkanowy/ui/modules/dashboard/DashboardPresenter.kt
@@ -5,7 +5,6 @@ import io.github.wulkanowy.data.dataOrNull
 import io.github.wulkanowy.data.db.entities.AdminMessage
 import io.github.wulkanowy.data.db.entities.LuckyNumber
 import io.github.wulkanowy.data.db.entities.Student
-import io.github.wulkanowy.data.db.entities.Timetable
 import io.github.wulkanowy.data.enums.MessageFolder
 import io.github.wulkanowy.data.enums.MessageType
 import io.github.wulkanowy.data.errorOrNull
@@ -24,13 +23,12 @@ import io.github.wulkanowy.data.repositories.SchoolAnnouncementRepository
 import io.github.wulkanowy.data.repositories.SemesterRepository
 import io.github.wulkanowy.data.repositories.StudentRepository
 import io.github.wulkanowy.data.repositories.TimetableRepository
-import io.github.wulkanowy.data.toFirstResult
 import io.github.wulkanowy.domain.adminmessage.GetAppropriateAdminMessageUseCase
+import io.github.wulkanowy.domain.timetable.IsStudentHasLessonsOnWeekendUseCase
 import io.github.wulkanowy.ui.base.BasePresenter
 import io.github.wulkanowy.ui.base.ErrorHandler
 import io.github.wulkanowy.utils.AdsHelper
 import io.github.wulkanowy.utils.calculatePercentage
-import io.github.wulkanowy.utils.monday
 import io.github.wulkanowy.utils.nextOrSameSchoolDay
 import io.github.wulkanowy.utils.sunday
 import kotlinx.coroutines.flow.Flow
@@ -47,7 +45,6 @@ import kotlinx.coroutines.flow.merge
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import timber.log.Timber
-import java.time.DayOfWeek
 import java.time.Instant
 import java.time.LocalDate
 import javax.inject.Inject
@@ -61,6 +58,7 @@ class DashboardPresenter @Inject constructor(
     private val messageRepository: MessageRepository,
     private val attendanceSummaryRepository: AttendanceSummaryRepository,
     private val timetableRepository: TimetableRepository,
+    private val isStudentHasLessonsOnWeekendUseCase: IsStudentHasLessonsOnWeekendUseCase,
     private val homeworkRepository: HomeworkRepository,
     private val examRepository: ExamRepository,
     private val conferenceRepository: ConferenceRepository,
@@ -440,16 +438,7 @@ class DashboardPresenter @Inject constructor(
     private fun loadLessons(student: Student, forceRefresh: Boolean) {
         flatResourceFlow {
             val semester = semesterRepository.getCurrentSemester(student)
-
-            val currentWeekLessons = timetableRepository.getTimetable(
-                student = student,
-                semester = semester,
-                start = LocalDate.now().monday,
-                end = LocalDate.now().sunday,
-                forceRefresh = false,
-            ).toFirstResult().dataOrNull?.lessons.orEmpty()
-
-            val date = when (isWeekendHasLessons(currentWeekLessons)) {
+            val date = when (isStudentHasLessonsOnWeekendUseCase(student, semester)) {
                 true -> LocalDate.now()
                 else -> LocalDate.now().nextOrSameSchoolDay
             }
@@ -458,8 +447,8 @@ class DashboardPresenter @Inject constructor(
                 student = student,
                 semester = semester,
                 start = date,
-                end = date.plusDays(1),
-                forceRefresh = forceRefresh
+                end = date.sunday,
+                forceRefresh = forceRefresh,
             )
         }
             .onEach {
@@ -494,15 +483,6 @@ class DashboardPresenter @Inject constructor(
                 }
             }
             .launchWithUniqueRefreshJob("dashboard_lessons", forceRefresh)
-    }
-
-    private fun isWeekendHasLessons(
-        lessons: List<Timetable>,
-    ): Boolean = lessons.any {
-        it.date.dayOfWeek in listOf(
-            DayOfWeek.SATURDAY,
-            DayOfWeek.SUNDAY,
-        )
     }
 
     private fun loadHomework(student: Student, forceRefresh: Boolean) {
